requirements.txt
playwright
PyYAML
pandas
pdfplumber
requests
sqlite3
tqdm
